removed sync point, double inst. of float

+++ src/esutil/CudaHelper.cuh
@@ -88,11 +88,11 @@ inline __host__ __device__ void operator+=(double3 &a, double3 b)
 {
     a.x += b.x; a.y += b.y; a.z += b.z;
 }
-inline __host__ __device__ double3 operator+(double3 a, float b)
+inline __host__ __device__ double3 operator+(double3 a, double b)
 {
     return make_double3(a.x + b, a.y + b, a.z + b);
 }
-inline __host__ __device__ void operator+=(double3 &a, float b)
+inline __host__ __device__ void operator+=(double3 &a, double b)
 {
     a.x += b; a.y += b; a.z += b;
 }
@@ -105,15 +105,15 @@ inline __host__ __device__ void operator+=(double4 &a, double4 b)
 {
     a.x += b.x; a.y += b.y; a.z += b.z; a.w += b.w;
 }
-inline __host__ __device__ double4 operator+(double4 a, float b)
+inline __host__ __device__ double4 operator+(double4 a, double b)
 {
     return make_double4(a.x + b, a.y + b, a.z + b, a.w + b);
 }
-inline __host__ __device__ double4 operator+(float b, double4 a)
+inline __host__ __device__ double4 operator+(double b, double4 a)
 {
     return make_double4(a.x + b, a.y + b, a.z + b, a.w + b);
 }
-inline __host__ __device__ void operator+=(double4 &a, float b)
+inline __host__ __device__ void operator+=(double4 &a, double b)
 {
     a.x += b; a.y += b; a.z += b; a.w += b;
 }
@@ -126,11 +126,11 @@ inline __host__ __device__ void operator-=(double4 &a, double4 b)
 {
     a.x -= b.x; a.y -= b.y; a.z -= b.z; a.w -= b.w;
 }
-inline __host__ __device__ double4 operator-(double4 a, float b)
+inline __host__ __device__ double4 operator-(double4 a, double b)
 {
     return make_double4(a.x - b, a.y - b, a.z - b,  a.w - b);
 }
-inline __host__ __device__ void operator-=(double4 &a, float b)
+inline __host__ __device__ void operator-=(double4 &a, double b)
 {
     a.x -= b; a.y -= b; a.z -= b; a.w -= b;
 }
@@ -143,15 +143,15 @@ inline __host__ __device__ void operator-=(double3 &a, double3 b)
 {
     a.x -= b.x; a.y -= b.y; a.z -= b.z;
 }
-inline __host__ __device__ double3 operator-(double3 a, float b)
+inline __host__ __device__ double3 operator-(double3 a, double b)
 {
     return make_double3(a.x - b, a.y - b, a.z - b);
 }
-inline __host__ __device__ double3 operator-(float b, double3 a)
+inline __host__ __device__ double3 operator-(double b, double3 a)
 {
     return make_double3(b - a.x, b - a.y, b - a.z);
 }
-inline __host__ __device__ void operator-=(double3 &a, float b)
+inline __host__ __device__ void operator-=(double3 &a, double b)
 {
     a.x -= b; a.y -= b; a.z -= b;
 }
@@ -164,15 +164,15 @@ inline __host__ __device__ void operator*=(double3 &a, double3 b)
 {
     a.x *= b.x; a.y *= b.y; a.z *= b.z;
 }
-inline __host__ __device__ double3 operator*(double3 a, float b)
+inline __host__ __device__ double3 operator*(double3 a, double b)
 {
     return make_double3(a.x * b, a.y * b, a.z * b);
 }
-inline __host__ __device__ double3 operator*(float b, double3 a)
+inline __host__ __device__ double3 operator*(double b, double3 a)
 {
     return make_double3(b * a.x, b * a.y, b * a.z);
 }
-inline __host__ __device__ void operator*=(double3 &a, float b)
+inline __host__ __device__ void operator*=(double3 &a, double b)
 {
     a.x *= b; a.y *= b; a.z *= b;
 }
@@ -185,24 +185,24 @@ inline __host__ __device__ void operator*=(double4 &a, double4 b)
 {
     a.x *= b.x; a.y *= b.y; a.z *= b.z; a.w *= b.w;
 }
-inline __host__ __device__ double4 operator*(double4 a, float b)
+inline __host__ __device__ double4 operator*(double4 a, double b)
 {
     return make_double4(a.x * b, a.y * b, a.z * b,  a.w * b);
 }
-inline __host__ __device__ double4 operator*(float b, double4 a)
+inline __host__ __device__ double4 operator*(double b, double4 a)
 {
     return make_double4(b * a.x, b * a.y, b * a.z, b * a.w);
 }
-inline __host__ __device__ void operator*=(double4 &a, float b)
+inline __host__ __device__ void operator*=(double4 &a, double b)
 {
     a.x *= b; a.y *= b; a.z *= b; a.w *= b;
 }
 
-inline __host__ __device__ float dot(double3 a, double3 b)
+inline __host__ __device__ double dot(double3 a, double3 b)
 { 
     return a.x * b.x + a.y * b.y + a.z * b.z;
 }
-inline __host__ __device__ float dot(double4 a, double4 b)
+inline __host__ __device__ double dot(double4 a, double4 b)
 { 
     return a.x * b.x + a.y * b.y + a.z * b.z + a.w * b.w;
 }
